Move updatePlayList to useSpotify hook
